realizado me sala

+++ filmes (1)/filmes/src/main/java/br/com/leonardo/filmes/repository/GeneroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.leonardo.filmes.model.Genero;
 
-public interface GeneroRepository extends JpaRepository<Genero, Long>{
+public interface GeneroRepository extends JpaRepository<Genero, Integer>{
 
 }
